backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
pandas>=2.0
openpyxl>=3.1
python-dotenv>=1.0
django-celery-beat>=2.5
celery>=5.3
django-filter>=23.5
reportlab>=4.0
